refactor: hook 함수를 분리합니다.
